train_RL/requirements.txt
diskcache==5.4.0
matplotlib==3.5.1
mmcv-full==1.5.2
numba==0.55.1
numpy==1.21.5
opencv_python==4.2.0.32
Pillow==9.4.0
scikit_image==0.19.2
setuptools==61.2.0
stable_baselines3==1.7.0
timm==0.6.12
torch==1.13.0
torch_scatter==2.1.0
torchvision==0.14.0
tqdm==4.64.0
ujson==5.1.0
tensorboard==2.11.2
wandb==0.13.5
mmsegmentation==0.25.0
mmdet==2.25.0
